added typescript to header.tsx
